特に経由する意味の無いSelectedFeedCollectionを削除してFormで行うように
